update main approvedraftorders tests
